consolidated stage resizing and content scale calculation code
The content scale factor now relies directly on the
`contentScaleFactor` property of the current view. Instead of assuming
the current view’s scale factor on start, a check is made every frame
and the back buffer is updated accordingly. The resizing and scale
factor are now only calculated when necessary and at the same time
every frame, thus making it easier to find any issues related with the
sizing or scale factor.